refactor: Sample controller to controller/sample package
